Fix if task name is too long line won't break Fixes #44
