+++ Frontend/Components/Screens/CategoriesScreen.tsx
@@ -1,11 +1,9 @@
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 import {View, StyleSheet, ScrollView, Dimensions} from 'react-native';
-import {Text, ListItem, Icon, Button} from '@rneui/themed';
+import {Text, Icon, Button} from '@rneui/themed';
 import Tasks from '../Tasks/Tasks';
-import {CategoryScreenProps, ITask} from '../../general/interfaces';
-import {useAppSelector} from '../../hooks/store';
+import {CategoryScreenProps} from '../../general/interfaces';
 import AddTaskDialog from '../Dialogs/AddTaskDialog';
-import {API_HOST} from '@env';
 
 const windowHeight = Dimensions.get('window').height;
 
@@ -13,33 +11,9 @@ export default function CategoriesScreen({
   category,
   isSubCategory,
 }: CategoryScreenProps): JSX.Element {
-  const [tasks, setTasks] = useState<ITask[]>([]);
   const [open, setOpen] = useState<boolean>(false);
   const [isUpdate, setIsUpdate] = useState<boolean>(false);
   const [iaEditMode, setIsEditMode] = useState<boolean>(false);
-
-  const userToken = useAppSelector(state => state.user.token);
-
-  useEffect(() => {
-    const fetchTasks = async () => {
-      try {
-        const url = isSubCategory
-          ? `SubCategory/${category?.id}/tasks`
-          : `Category/${category?.id}/tasks`;
-        const response = await fetch(`${API_HOST}api/` + url, {
-          headers: {
-            Authorization: `Bearer ${userToken}`,
-          },
-        });
-        const data = await response.json();
-        data.reverse();
-        setTasks(data);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    fetchTasks();
-  }, [userToken, category, isSubCategory, isUpdate]);
 
   return (
     <View style={styles.container}>
@@ -62,25 +36,13 @@ export default function CategoriesScreen({
         />
       </View>
       <ScrollView style={styles.scrollView}>
-        {Array.isArray(tasks) && tasks.length > 0 && tasks ? (
-          <Tasks
-            tasks={tasks}
-            category={category}
-            isSubCategory={isSubCategory}
-            isEditMode={iaEditMode}
-            onUpdate={() => setIsUpdate(!isUpdate)}
-          />
-        ) : (
-          <>
-            <ListItem containerStyle={styles.listItemContainer}>
-              <ListItem.Content>
-                <ListItem.Title>
-                  <Text style={styles.listTitle}>No tasks yet</Text>
-                </ListItem.Title>
-              </ListItem.Content>
-            </ListItem>
-          </>
-        )}
+        <Tasks
+          isUpdate={isUpdate}
+          category={category}
+          isSubCategory={isSubCategory}
+          isEditMode={iaEditMode}
+          onUpdate={() => setIsUpdate(!isUpdate)}
+        />
         <AddTaskDialog
           open={open}
           onBackPress={() => setOpen(false)}
@@ -102,13 +64,6 @@ const styles = StyleSheet.create({
     paddingTop: '10%',
     paddingHorizontal: '10%',
   },
-  listItemContainer: {
-    flex: 1,
-    flexDirection: 'column',
-    backgroundColor: '#262c2e',
-    borderBottomColor: 'white',
-    borderBottomWidth: 1,
-  },
   list: {
     flexGrow: 1,
     paddingTop: '10%',
@@ -120,11 +75,6 @@ const styles = StyleSheet.create({
   scrollView: {
     marginTop: '10%',
     marginBottom: '20%',
-  },
-  listTitle: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: 'bold',
   },
   iconscontainer: {
     flexDirection: 'row',
